Create line graph to show trend in cases/deaths
react-chartjs-2 library is used to create a line graph which maps through each object in the state array and uses each reportDate as a label, with confirmed.total and deaths.total as the datasets. A check is made on the state first to see if the fetchApi function has returned before mapping through it and rendering the chart.