Create bar chart to display stats from selected countries
A bar chart is created and takes the data from App.js props to populate a bar chart. The return statement checks to see if the country property exists in the App.js state which would indicate that a user has made a selection, then return a bar chart if so. Otherwise, the line chart for the global stats will be rendered